Moved the assistant streaming endpoints.
