+++ settings.py
@@ -49,3 +49,6 @@ BLOOMS_LLM_ENDPOINT = os.environ.get('BLOOMS_LLM_ENDPOINT', 'http://localhost:50
 
 OPENPROJECT_API_KEY = os.environ.get("OPENPROJECT_API_KEY")
 OPENPROJECT_URL = os.environ.get("OPENPROJECT_URL")
+
+
+DEFAULT_ASSISTANT_ID = "asst_X0dIT6aOTHFQgJNE923sjv8E"
